+++ tests/testthat/test_micomp.R
@@ -91,6 +91,20 @@ test_that("micomp constructs the expected objects", {
                          files = c(files, files),
                          lvls = c("NLOK", "JEXOK"))))
 
+  # 5 - One named output, one variance to explain, two comparisons, explicit
+  # Pillay test
+  mic5 <- micomp("TheOutput", 0.9,
+                  list(
+                    list(name = "NLOKvsJEXNOSHUFF",
+                         folders = c(dir_nl_ok, dir_jex_noshuff),
+                         files = c(files, files),
+                         lvls = c("NLOK", "JEXNOSHUFF")),
+                    list(name = "NLOKvsJEXDIFF",
+                         folders = c(dir_nl_ok, dir_jex_diff),
+                         files = c(files, files),
+                         lvls = c("NLOK", "JEXDIFF"))),
+                  mnv_test = "Pillai")
+
   ##### Start testing #####
 
   # Check object dimensions
@@ -99,6 +113,7 @@ test_that("micomp constructs the expected objects", {
   expect_equal(dim(mic2), c(7, 3))
   expect_equal(dim(mic3), c(6, 3))
   expect_equal(dim(mic4), c(1, 1))
+  expect_equal(dim(mic5), c(1, 2))
 
   # Check object row names
   expect_equal(rownames(mic1a), outputs)
@@ -107,6 +122,7 @@ test_that("micomp constructs the expected objects", {
   expect_equal(rownames(mic3),
                c("out1", "out2", "out3", "out4", "out5", "out6"))
   expect_equal(rownames(mic4), c("out1"))
+  expect_equal(rownames(mic5), c("TheOutput"))
 
   # Check object column names
   expect_equal(colnames(mic1a),
@@ -118,6 +134,7 @@ test_that("micomp constructs the expected objects", {
   expect_equal(colnames(mic3),
                c("NLOKvsJEXOK", "NLOKvsJEXNOSHUFF", "NLOKvsJEXDIFF"))
   expect_equal(colnames(mic4), "NLOKvsJEXOK")
+  expect_equal(colnames(mic5), c("NLOKvsJEXNOSHUFF", "NLOKvsJEXDIFF"))
 
   # Check properties of sub-objects
   for (i in 1:dim(mic1a)[1]) {