+++ NEWS.md
@@ -1,6 +1,7 @@
 # micompr 1.0.1
 
 * Fix bug in grpoutputs when loading files with just one output
+* Fix bug in micomp when using one named output
 
 # micompr 1.0.0